user get and delete added

+++ Backend/flask_app/main/service/user_service.py
@@ -1,4 +1,4 @@
-import uuid
+#import uuid
 #import datetime
 
 from .. import db
@@ -11,6 +11,21 @@ def get_all_users():
             'data': [user.toDict() for user in users],
         }
     return response_object, 200
+
+def get_user(id):
+    user = User.query.get(id)
+    if user:
+        response_object = {
+            'status': 'success',
+            'data': [user.toDict()],
+        }
+        return response_object, 200
+    else:
+        response_object = {
+            'status': 'fail',
+            'message': 'User not found.'
+        }
+        return response_object, 404
 
 def save_new_user(data):
     user = User.query.filter_by(email=data['email']).first()
@@ -38,14 +53,21 @@ def save_new_user(data):
         return response_object, 409
 
 def delete_user(id):
-    user = User.query.get_or_404(id)
-    db.session.delete(user)
-    db.session.commit()
-    response_object = {
-           'status': 'success',
-           'message': 'Successfully deleted.'
+    user = User.query.get(id)
+    if user:
+        db.session.delete(user)
+        db.session.commit()
+        response_object = {
+            'status': 'success',
+            'message': 'Successfully deleted.'
         }
-    return response_object, 201
+        return response_object, 200
+    else:
+        response_object = {
+            'status': 'fail',
+            'message': 'User not found.'
+        }
+        return response_object, 404
 
 def save_changes(data):
     db.session.add(data)
